# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Trading Simulator API",
    "version": "1.0.0",
    "description": "API for the Trading Simulator - a platform for simulated cryptocurrency trading competitions\n      \n## Authentication Guide\n\nThis API uses Bearer token authentication. All protected endpoints require the following header:\n\n- **Authorization**: Bearer your-api-key\n\nWhere \"your-api-key\" is the API key provided during team registration.\n\n### Authentication Examples\n\n**cURL Example:**\n\n```bash\ncurl -X GET \"https://api.example.com/api/account/balances\" \\\n  -H \"Authorization: Bearer abc123def456_ghi789jkl012\" \\\n  -H \"Content-Type: application/json\"\n```\n\n**JavaScript Example:**\n\n```javascript\nconst fetchData = async () => {\n  const apiKey = 'abc123def456_ghi789jkl012';\n  const response = await fetch('https://api.example.com/api/account/balances', {\n    headers: {\n      'Authorization': `Bearer ${apiKey}`,\n      'Content-Type': 'application/json'\n    }\n  });\n  \n  return await response.json();\n};\n```\n\nFor convenience, we provide an API client that handles authentication automatically. See `docs/examples/api-client.ts`.\n      ",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "ISC License",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
...
    },
    {
...
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key provided in the Authorization header using Bearer token authentication"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of when the error occurred"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/price": {
      "get": {
        "tags": [
          "Price"
        ],
        "summary": "Get price for a token",
        "description": "Get the current price of a specified token",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Bearer token for authentication (format \"Bearer YOUR_API_KEY\")",
            "example": "Bearer abc123def456_ghi789jkl012"
          },
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Token address",
            "example": "So11111111111111111111111111111111111111112"
          },
          {
            "in": "query",
            "name": "chain",
            "schema": {
              "type": "string",
              "enum": [
                "evm",
                "svm"
              ]
            },
            "required": false,
            "description": "Blockchain type of the token",
            "example": "svm"
          },
          {
            "in": "query",
            "name": "specificChain",
            "schema": {
              "type": "string",
              "enum": [
                "eth",
                "polygon",
                "bsc",
                "arbitrum",
                "optimism",
                "avalanche",
                "base",
                "linea",
                "zksync",
                "scroll",
                "mantle",
                "svm"
              ]
            },
            "required": false,
            "description": "Specific chain for EVM tokens",
            "example": "eth"
          }
        ],
        "responses": {
          "200": {
            "description": "Token price information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Whether the price was successfully retrieved"
                    },
                    "price": {
                      "type": "number",
                      "nullable": true,
                      "description": "Current price of the token in USD"
                    },
                    "token": {
                      "type": "string",
                      "description": "Token address"
                    },
                    "chain": {
                      "type": "string",
                      "enum": [
                        "EVM",
                        "SVM"
                      ],
                      "description": "Blockchain type of the token"
                    },
                    "specificChain": {
                      "type": "string",
                      "nullable": true,
                      "description": "Specific chain for EVM tokens"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Timestamp when the price was fetched"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/price/token-info": {
      "get": {
        "tags": [
          "Price"
        ],
        "summary": "Get detailed token information",
        "description": "Get detailed token information including price and specific chain",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Bearer token for authentication (format \"Bearer YOUR_API_KEY\")",
            "example": "Bearer abc123def456_ghi789jkl012"
          },
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Token address",
            "example": "So11111111111111111111111111111111111111112"
          },
          {
            "in": "query",
            "name": "chain",
            "schema": {
              "type": "string",
              "enum": [
                "evm",
                "svm"
              ]
            },
            "required": false,
            "description": "Blockchain type of the token",
            "example": "svm"
          },
          {
            "in": "query",
            "name": "specificChain",
            "schema": {
              "type": "string",
              "enum": [
                "eth",
                "polygon",
                "bsc",
                "arbitrum",
                "optimism",
                "avalanche",
                "base",
                "linea",
                "zksync",
                "scroll",
                "mantle",
                "svm"
              ]
            },
            "required": false,
            "description": "Specific chain for EVM tokens",
            "example": "eth"
          }
        ],
        "responses": {
          "200": {
            "description": "Token information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Whether the token information was successfully retrieved"
                    },
                    "price": {
                      "type": "number",
                      "nullable": true,
                      "description": "Current price of the token in USD"
                    },
                    "token": {
                      "type": "string",
                      "description": "Token address"
                    },
                    "chain": {
                      "type": "string",
                      "enum": [
                        "EVM",
                        "SVM"
                      ],
                      "description": "Blockchain type of the token"
                    },
                    "specificChain": {
                      "type": "string",
                      "nullable": true,
                      "description": "Specific chain for EVM tokens"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid authentication"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}